requirements.txt
Django==1.9
line-bot-sdk==1.5.0
gunicorn==19.3.0
jieba==0.39
gensim==3.8.0
six==1.12.0
django-redis==4.10.0
redis==3.3.8
django-redis-cache==2.1.0